+++ src/main/java/GaonNuri/Project/ShoppingMall/admin/service/inter/AdminService.java
@@ -1,0 +1,10 @@
+package GaonNuri.Project.ShoppingMall.admin.service.inter;
+
+import GaonNuri.Project.ShoppingMall.admin.dto.ItemsRequestDto;
+
+public interface AdminService {
+
+    void updateItemsInfo(ItemsRequestDto.ItemsUpdateInfo dto);
+
+    void registerItemsInfo(ItemsRequestDto.ItemsRegisterInfo dto);
+}
