+++ src/main/java/GaonNuri/Project/ShoppingMall/item/controller/ItemsController.java
@@ -21,7 +21,6 @@ public class ItemsController {
     //상품 목록페이지
     @GetMapping()
     public Page<ItemsInfo> getAllItems(@RequestParam("page") Integer page, @RequestParam("size") Integer size){
-
         return itemService.showItemsOnly(page, size);
     }
