+++ src/main/java/GaonNuri/Project/ShoppingMall/item/data/entity/Items.java
@@ -6,6 +6,8 @@ import lombok.*;
 
 import javax.persistence.*;
 
+import static GaonNuri.Project.ShoppingMall.admin.dto.ItemsRequestDto.ItemsUpdateInfo;
+
 @Entity
 @Getter
 @Builder
@@ -32,5 +34,16 @@ public class Items extends BaseTimeEntity {
 
     @Enumerated(EnumType.STRING)
     private ItemStatus itemStatus;
+
+    public void updateItem(ItemsUpdateInfo dto){
+        if(dto.getItemName()!=null)
+            this.itemName = dto.getItemName();
+        if(dto.getPrice()>=0)
+            this.price = dto.getPrice();
+        if(dto.getItemDetail()!=null)
+            this.itemDetail = dto.getItemDetail();
+        if(dto.getItemStatus()!=null)
+            this.itemStatus = dto.getItemStatus();
+    }
 
 }
