+++ src/main/java/GaonNuri/Project/ShoppingMall/member/data/dto/EntryPointErrorResponse.java
@@ -1,0 +1,15 @@
+package GaonNuri.Project.ShoppingMall.member.data.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@ToString
+public class EntryPointErrorResponse {
+
+    private String msg;
+}
