+++ src/main/java/GaonNuri/Project/ShoppingMall/member/data/dto/TokenDto.java
@@ -1,0 +1,18 @@
+package GaonNuri.Project.ShoppingMall.member.data.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class TokenDto {
+
+    private String grantType;
+    private String accessToken;
+    private String refreshToken;
+    private Long accessTokenExpiresIn;
+}
