+++ src/main/java/GaonNuri/Project/ShoppingMall/member/data/dto/TokenRequestDto.java
@@ -1,0 +1,11 @@
+package GaonNuri.Project.ShoppingMall.member.data.dto;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class TokenRequestDto {
+    private String accessToken;
+    private String refreshToken;
+}
