+++ src/main/java/GaonNuri/Project/ShoppingMall/member/data/entity/RefreshToken.java
@@ -1,0 +1,34 @@
+package GaonNuri.Project.ShoppingMall.member.data.entity;
+
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Entity
+public class RefreshToken {
+
+    @Id
+    @Column(name = "rt_key")
+    private String key;
+
+    @Column(name = "rt_value")
+    private String value;
+
+    @Builder
+    public RefreshToken(String key, String value) {
+        this.key = key;
+        this.value = value;
+    }
+
+    public RefreshToken updateValue(String token){
+        this.value = token;
+        return this;
+    }
+}
