+++ src/main/java/GaonNuri/Project/ShoppingMall/member/data/enums/AuthorityEnum.java
@@ -1,0 +1,5 @@
+package GaonNuri.Project.ShoppingMall.member.data.enums;
+
+public enum AuthorityEnum {
+    ROLE_USER, ROLE_ADMIN
+}
