+++ src/main/java/GaonNuri/Project/ShoppingMall/member/data/validation/LoginValidation.java
@@ -1,0 +1,4 @@
+package GaonNuri.Project.ShoppingMall.member.data.validation;
+
+public interface LoginValidation {
+}
