+++ src/main/java/GaonNuri/Project/ShoppingMall/member/repository/MemberRepository.java
@@ -1,0 +1,11 @@
+package GaonNuri.Project.ShoppingMall.member.repository;
+
+import GaonNuri.Project.ShoppingMall.member.data.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+    Optional<Member> findByEmail(String email);
+    boolean existsByEmail(String email);
+}
