--- src/main/java/GaonNuri/Project/ShoppingMall/user/data/dto/EntryPointErrorResponse.java
@@ -1,15 +1,0 @@
-package GaonNuri.Project.ShoppingMall.user.data.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@ToString
-public class EntryPointErrorResponse {
-
-    private String msg;
-}
