--- src/main/java/GaonNuri/Project/ShoppingMall/user/data/dto/TokenRequestDto.java
@@ -1,11 +1,0 @@
-package GaonNuri.Project.ShoppingMall.user.data.dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class TokenRequestDto {
-    private String accessToken;
-    private String refreshToken;
-}
