--- src/main/java/GaonNuri/Project/ShoppingMall/user/data/enums/AuthorityEnum.java
@@ -1,5 +1,0 @@
-package GaonNuri.Project.ShoppingMall.user.data.enums;
-
-public enum AuthorityEnum {
-    ROLE_USER, ROLE_ADMIN
-}
