--- src/main/java/GaonNuri/Project/ShoppingMall/user/data/validation/JoinValidation.java
@@ -1,4 +1,0 @@
-package GaonNuri.Project.ShoppingMall.user.data.validation;
-
-public interface JoinValidation {
-}
