--- src/main/java/GaonNuri/Project/ShoppingMall/user/data/validation/LoginValidation.java
@@ -1,4 +1,0 @@
-package GaonNuri.Project.ShoppingMall.user.data.validation;
-
-public interface LoginValidation {
-}
