--- src/main/java/GaonNuri/Project/ShoppingMall/user/data/validation/UpdateValidation.java
@@ -1,4 +1,0 @@
-package GaonNuri.Project.ShoppingMall.user.data.validation;
-
-public interface UpdateValidation {
-}
